Divide type traits into separate files
